refactor(shared/ui): shadcn 컴포넌트 디렉토리 분리
`shared/ui`에 혼재되어 있던 shadcn/ui 컴포넌트와 직접 구현 컴포넌트를 분리하고,
mock 관련 폴더명을 `mock`으로 통일합니다.

**shared/ui 구조 분리**
- `shared/ui/shadcn/` 디렉토리 신설 및 shadcn 컴포넌트 이동
  - `avatar`, `button`, `input`, `separator`, `sheet`, `sidebar`, `skeleton`, `tooltip`
- 직접 구현 컴포넌트(`Logo.tsx`)는 `shared/ui/` 루트에 유지
- shadcn 컴포넌트 import 시 `index.ts` 경유 없이 직접 경로 사용
  - e.g. `import { Button } from '@/shared/ui/shadcn/button'`
- `components.json` alias 수정 → 이후 `npx shadcn add` 시 `shadcn/` 경로에 자동 설치
