+++ components.json
@@ -13,9 +13,9 @@
   "iconLibrary": "lucide",
   "rtl": false,
   "aliases": {
-    "components": "@/shared/ui",
+    "components": "@/shared/ui/shadcn",
     "utils": "@/shared/lib/utils",
-    "ui": "@/shared/ui",
+    "ui": "@/shared/ui/shadcn",
     "lib": "@/shared/lib",
     "hooks": "@/shared/hooks"
   },
