+++ src/features/navigation/ui/NavLockedItem.tsx
@@ -1,4 +1,4 @@
-import { SidebarMenuItem, SidebarMenuButton } from '@/shared/ui/sidebar'
+import { SidebarMenuItem, SidebarMenuButton } from '@/shared/ui/shadcn/sidebar'
 import { SidebarIcon } from '@/features/navigation/ui/NavSidebarIcon'
 import type { NavItem } from '../model/types'
 import Link from 'next/link'
