+++ src/shared/ui/shadcn/avatar.tsx
